Adds initial `secrets` command processor
https://www.pivotaltracker.com/story/show/161188859

This change adds the beginning of the `secrets` command. It currently
is supporting the `keypair` sub-command with the `--generate` flag. So:
`state secrets keypair --generate`.

This change renames `api.bearerToken` to `api.BearerToken`; thus making it a
public variable.

This change adds a test helper (`internal/testhelpers/osutil`) which
makes it easier to execute a function and capture anything written to
stdout. At somepoint, it should just capture anything written to stdout
and stderr; just don't need stderr yet.