Adds secret sharing w/specific user
https://www.pivotaltracker.com/story/show/161315404

This change makes it possible to share non-user scoped secrets with
another user in the same organization. For example:

```
state secrets share scottr
```

Will share org and project scoped secrets with `scottr` by using that
user's public key to encrypt those secrets and upload them to the
secrets service on scottr's behalf. This basically touches everything in
the secrets service to date, sans deleting a secret.

Additionally, adds missing tests for:

- api.ErrorCode
- keypair.FetchX

... fixes an issue with the Makefile where I had stupidly ignore any
packages with the tri-gram `api` in them. Whoopsie.