Allows api settings to be defined in preprocessor
https://www.pivotaltracker.com/story/show/162270594

This change allows API and SecretsAPI settings to be defined in the
preprocessor, eliminating the need to decide anything at runtime over
what env settings to use. It also introduces an env settings map in the
preprocessor which can be extended by simply adding more.

Code has been updated in the CLI to simply always use the constants.