Requires `secrets` commands to load pk form local file
https://www.pivotaltracker.com/story/show/161922416

This change makes it so that `state secrets` commands that need to
expand a secret for any reason use an unecrypted pk file from the config
dir instead of trying to download the full keypair from secrets service.
This way, the user does not need to enter their passphrase each time.

An error message will suggest the user run `state keypair auth` if there
are issues trying to load the local pk file.