Obtain GOPATH and GOROOT from go if they are not set
If the user has not explicitely set GOPATH or GOROOT, we extract the
implicit values from the go binary in the user's path.  When running in
`state activate` this is not an issue, since the state tool is setting
these values for us, but there's a bootstrapping problem for a user
trying to build the tool without having the tool in an environment where
GOPATH and GOROOT are not explicitly set.