Stops checking if we're in a test in the state auth code
https://www.pivotaltracker.com/story/show/163197968

This change removes the logical branches from the `state auth` commands
which were concerned with whether or not a unit test was running. This
change also removed the `testX` vars from the same package. Instead, the
tests are now negotiating answers to prompts via stdin buffering.

I had to do some investigation to sort out the right way to write into
stdin and have the `survey` package work correctly with prompts. This is
why the write to stdin with delays code was added.