Renames "commands" to "scripts" in activestate.yaml parser
https://www.pivotaltracker.com/story/show/163167327

This change only addresses the field in the AS project file (i.e.
activestate.yaml) named "commands". It has been renamed to "scripts" and
all tests updated accordingly.

This change does NOT rename the `internal/projectfile/Project.Commands`
field to `Scripts`. Thus, all code in this change still refers to
scripts as Commands.