Moves APY installer to runtime and redefines installer.Installer
This change converts the ActivePythonInstaller to a
platform/runtime.Installer and redefines the installer.Installer to be
something much more generic. This way, I can define a composite
Installer than can accept a runtime.Downloader and runtime.Installer
which can then be used back in the virtualenvironment package. This
composite Installer is the installer.RuntimeInstaller.

Also, this defines the runtime.Downloader interface.