refactor platform/runtime to support alternative builder
This commit splits out all the builder specific logic from the
installation.

It should not change any of the existing behavior, just re-organize it.

The only significant change is in pkg/platform/runtime/download.go,
where we are now parsing for the build_engine field in the build request
response.

If the build_engine is not `camel`, an error is  returned in the
installer method.