support --output mono flag
This commit adds a global option `--output mono` that forces monochrome
output.

If the environemnt variable `TERM` is set to `dumb`, the output will
also be monochrome.

Futhermore, we check if `stdout` has terminal capabilities, and if not,
for example because the output is redirected to a file, colors are
disabled.