also check complete installation for alternative builds
This commit introduces a marker file that is created after an
installation is completed.

Only alternative builds currently check whether this file exists, and
return an error message advising to delete the installation directory.
Camel builds are not included, as all cached installations would trigger
this behavior, which is considered backwards incompatible.