relocate script shebangs in binary directory
The only part of a Python MacOS build that needs to be relocated are the
script files in the binary directory.

This commit adds functionality to do targeted relocations, and
configures a targeted relocation for the specific use case.

The activation integration tests for Python2 and Python3 are updated to
test whether installed scripts (for `pip2` and `pip3`) are functional.