hotfix: commitID is always set
This is a hotfix after story: https://www.pivotaltracker.com/story/show/172863539

The integration test for `state export` failed, because of an unset
commitID.

It also revealed that the API response for a project does not respond
the Organization in its name(!) or any other field.  So, that had to be
fixed too.  I thought I checked that before... Ooops!