base version for CI version incrementer is on S3
When incrementing the version on CI, we download the base version from
S3.  From the same spot that the install.sh and install.ps1 scripts get
their version and artifact information from.

To rely on a single source of truth, we are using the linux-amd64.json
file independent of the architecture that we are upgrading for.