consistent uninstall behavior
If an error occurs at any point, we are still trying to remove as much
as possible.

We also report the files that could not be removed, such that the user
can remove them manually.  That appears to be a better approach than
leaving the user with a half-uninstallation, and no way to fix it or
re-try it.

An even better choice might be a functionality to rollback the
uninstallation if anything fails: https://www.pivotaltracker.com/story/show/178134918