Use functions for state script instead of aliases
Using aliases in bash to add state tool scripts to the virtual
environments has a number of disadvantage.  One major drawback is that
these aliases are not available in subshells, or available to scripts.

That means doing things like:
```
$ myvar=$(mystatescript)
```
don't work.

Instead, create shell functions and export them with `export -f` so that
they can be used everywhere a regular command could be.