Ensure that artifact built events are not counted twice
After talking to Stephen Reichling, I realized that the `cache_hit`
field in the artifact progress messages only indicates that an artifact
has been built as part of a different recipe before.

It does not, however, check if the artifact has to be built for this
specific request.  Hence, I had to get that information from the
headchef build result.