Remove user env before deployment.
On Windows in particular, it's easy to get the "Unable to remove deployed runtime" error if the directory is in use. It is often the case that all but the top-level directory is removed. Clean the user env first since attempting to run the command again will result in a "There is no deployed runtime" error.